Auth+error handling

+++ backend/src/main/java/com/java55/itvitaetodolist/list/ToDoListController.java
@@ -29,12 +29,18 @@ public class ToDoListController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ToDoListDto> get(@PathVariable long id){
+    public ResponseEntity<ToDoListDto> get(@PathVariable long id, Authentication authentication){
         var possibleList = toDoListService.findById(id);
         if(possibleList.isEmpty()){
             throw new NotFoundException();
         }
-        return ResponseEntity.ok(ToDoListDto.from(possibleList.get()));
+        User currentUser = (User) authentication.getPrincipal();
+        var foundList = possibleList.get();
+
+        if(!foundList.hasUser(currentUser)){
+            throw new NotFoundException();
+        }
+        return ResponseEntity.ok(ToDoListDto.from(foundList));
     }
 
     @PostMapping
@@ -53,7 +59,7 @@ public class ToDoListController {
     }
 
     @PatchMapping("{id}")
-    public ResponseEntity<?> update(@RequestBody ToDoListPatchDto toDoListPatchDto, @PathVariable Long id){
+    public ResponseEntity<?> update(@RequestBody ToDoListPatchDto toDoListPatchDto, @PathVariable Long id, Authentication authentication){
         var idFromBody = toDoListPatchDto.id();
         if(idFromBody != null){
             throw new BadRequestException("PATCH should not have id in body");
@@ -65,6 +71,12 @@ public class ToDoListController {
         }
         var originalToDoList = possibleOriginal.get();
 
+        var currentUser = (User) authentication.getPrincipal();
+
+        if(!originalToDoList.hasUser(currentUser)){
+            throw new NotFoundException();
+        }
+
         if(toDoListPatchDto.name() != null){
             originalToDoList.setName(toDoListPatchDto.name());
         }
@@ -74,13 +86,18 @@ public class ToDoListController {
         return ResponseEntity.ok(ToDoListDto.from(originalToDoList));
     }
 
-    @PostMapping("{id}")
-    public ResponseEntity<ToDoListDto> addUser(@PathVariable Long id, @RequestBody UserDto newUser){
+    @PostMapping("{id}/users")
+    public ResponseEntity<ToDoListDto> addUser(@PathVariable Long id, @RequestBody UserDto newUser, Authentication authentication){
         var possibleList = toDoListService.findById(id);
         if(possibleList.isEmpty()){
             throw new NotFoundException();
         }
         var list = possibleList.get();
+        var currentUser = (User) authentication.getPrincipal();
+
+        if(!list.hasUser(currentUser)){
+            throw new NotFoundException();
+        }
 
         if(newUser.username() == null || newUser.username().isBlank()){
             throw new BadRequestException("username can't be empty");
@@ -92,13 +109,19 @@ public class ToDoListController {
         return ResponseEntity.ok(ToDoListDto.from(list));
     }
 
-    @DeleteMapping("{id}/{username}")
-    public ResponseEntity<ToDoListDto> removeUser(@PathVariable Long id, @PathVariable String username){
+    @DeleteMapping("{id}/users/{username}")
+    public ResponseEntity<ToDoListDto> removeUser(@PathVariable Long id, @PathVariable String username, Authentication authentication){
         var possibleList = toDoListService.findById(id);
         if(possibleList.isEmpty()){
             throw new NotFoundException();
         }
         var list = possibleList.get();
+
+        var currentUser = (User) authentication.getPrincipal();
+
+        if(!list.hasUser(currentUser)){
+            throw new NotFoundException();
+        }
 
         toDoListService.removeUser(list, username);
 
@@ -106,15 +129,22 @@ public class ToDoListController {
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<Void> deleteList(@PathVariable long id){
+    public ResponseEntity<Void> deleteList(@PathVariable long id, Authentication authentication){
         var possibleList = toDoListService.findById(id);
         if(possibleList.isEmpty()){
             throw new NotFoundException();
         }
         var listToDelete = possibleList.get();
+
+        var currentUser = (User) authentication.getPrincipal();
+
+        if(!listToDelete.hasUser(currentUser)){
+            throw new NotFoundException();
+        }
         listToDelete.setEnabled(false);
         toDoListService.save(listToDelete);
 
         return ResponseEntity.noContent().build();
     }
+
 }
